requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bayesian-optimization==1.3.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cma==3.2.2
cmake==3.26.3
comm==0.1.3
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
docker-pycreds==0.4.0
dpcpp-cpp-rt==2022.2.1
executing==1.2.0
filelock==3.8.0
fonttools==4.29.1
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.11.1
idna==3.3
intel-cmplr-lib-rt==2022.2.1
intel-cmplr-lic-rt==2022.2.1
intel-opencl-rt==2022.2.1
intel-openmp==2022.2.1
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lightning-utilities==0.8.0
lit==16.0.5
llvmlite==0.40.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl==2022.2.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
munkres==1.1.4
nest-asyncio==1.5.6
networkx==3.1
nevergrad==0.5.0
numba==0.57.1
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==21.3
pandas==2.0.3
parso==0.8.3
pathlib==1.0.1
pathos==0.3.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
pox==0.3.2
ppft==1.7.6.6
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pydicom==2.3.1
Pygments==2.15.1
pynndescent==0.5.10
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2021.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.27.1
safetensors==0.3.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
sentry-sdk==1.5.7
setproctitle==1.3.2
shortuuid==1.0.8
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
stack-data==0.6.2
sympy==1.12
tbb==2021.7.1
termcolor==1.1.0
termcolor-whl==1.1.2
threadpoolctl==3.1.0
timm==0.9.2
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.11.4
torchvision==0.14.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
umap-learn==0.5.3
urllib3==1.26.8
wandb==0.15.3
wcwidth==0.2.6
yarl==1.9.2
yaspin==2.1.0
